Revert "feat: move query performance polling to its own celery task (#22372)"
This reverts commit a8e9fc890aaf171094505349d30f29aaaaf3e638.
