+++ frontend/src/scenes/insights/EmptyStates/EmptyStates.tsx
@@ -118,7 +118,6 @@ export function InsightLoadingStateWithLoadingBar({
         return () => clearInterval(interval)
     }, [insightPollResponse])
     const bytesPerSecond = (bytesRead / (secondsElapsed || 1)) * 1000
-    const estimatedRows = insightPollResponse?.status?.query_progress?.estimated_rows_total
 
     const cpuUtilization =
         (insightPollResponse?.status?.query_progress?.active_cpu_time || 0) /
@@ -133,11 +132,7 @@ export function InsightLoadingStateWithLoadingBar({
                 <p className="mx-auto text-center text-xs">
                     {rowsRead > 0 && bytesRead > 0 && (
                         <>
-                            {humanFriendlyNumber(rowsRead || 0, 0)}{' '}
-                            {estimatedRows && estimatedRows >= rowsRead
-                                ? `/ ${humanFriendlyNumber(estimatedRows)} `
-                                : null}{' '}
-                            rows
+                            {humanFriendlyNumber(rowsRead || 0, 0)} rows
                             <br />
                             {humanFileSize(bytesRead || 0)} ({humanFileSize(bytesPerSecond || 0)}/s)
                             <br />
